pyproject.toml
[tool.poetry]
name = "pdf-qa"
version = "0.2.0"
description = ""
authors = ["Robert McDermott"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<3.12"
qdrant-client = "^1.5.4"
pypdf2 = "^3.0.1"
langchain = "^0.0.300"
streamlit = "^1.27.0"
tiktoken = "^0.5.1"
python-dotenv = "^1.0.0"
openai = "^0.28.0"
streamlit-option-menu = "^0.3.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
